requirements.txt
Django>=4.1.7,<4.2
djangorestframework>=3.14,<3.15
psycopg2>=2.9.5,<2.10
Pillow>=9.4,<9.5
